First steps towards a CL lib to read Postgres dump file.
The idea is to allow pgloader to read Postgres dump files as a database
source of sorts. That would allow multiple things to happen when loading
data from a Postgres dump:

  - restore in parallel even with a custom dump file
  - benefit from pgloader smart parallel index creation
  - target a different schema name or table name than the source one
  - change other properties such as column names or types etc…

This first patch only implements reading the Table Of Contents from a
Postgres dump. That's quite something already, in its own way.